refactor: make the crop scan and diagnose endpoint resourceful - move the route from `farmer_profile_routes` to `crop_scan_routes` - move the controller method from `farmer_profiles_controller.diagnose` into `crop_scans.store` - move the test file accordingly - update the docs in README
