+++ app/controllers/farmer_profiles_controller.ts
@@ -3,13 +3,8 @@ import type { HttpContext } from '@adonisjs/core/http'
 import { schema } from '@adonisjs/validator'
 import hash from '@adonisjs/core/services/hash'
 import db from '@adonisjs/lucid/services/db'
-import fs from 'node:fs/promises'
-import AiService, { AIDiagnosis } from '#services/ai_service'
 import { rules } from '#services/validator_rules'
 import router from '@adonisjs/core/services/router'
-import CropScan from '#models/crop_scan'
-import { ProductCategory } from '#models/product'
-import { getCropTreatmentResults } from '../../helpers/crop_scan_helper.js'
 
 export default class FarmerProfilesController {
   /**
@@ -152,103 +147,9 @@ export default class FarmerProfilesController {
         phone_number: user.phone_number,
         role: user.role,
         farmerProfile: user.farmerProfile,
-      },
-    })
-  }
-
-  /**
-   * Scan a crop and get diagnosis and treatment results.
-   *
-   * `POST /api/v1/farmer_profiles/:farmer_profile_id/diagnose`
-   */
-  public async diagnose({ request, response, auth }: HttpContext) {
-    const user = auth.user!
-
-    const maxSize = '10mb'
-    const supportedExtNames = ['jpg', 'jpeg', 'png', 'webp', 'heic', 'heif']
-
-    const { image } = await request.validate({
-      schema: schema.create({
-        image: schema.file({
-          size: maxSize,
-          extnames: supportedExtNames,
-        }),
-      }),
-      messages: {
-        'image.required': 'Image is required.',
-        'image.file.size': `Image size must not exceed ${maxSize}.`,
-        'image.file.extname': `Image extension is not supported. Only ${supportedExtNames.join(', ')} are supported.`,
-      },
-    })
-
-    if (!image.tmpPath) {
-      return response.badRequest({ error: 'File upload failed' })
-    }
-
-    const imageBuffer = await fs.readFile(image.tmpPath)
-
-    const mimeType = image.extname === 'jpg' ? 'image/jpeg' : `image/${image.extname}`
-
-    let aiResult: AIDiagnosis | null = null
-
-    try {
-      aiResult = await AiService.diagnose(imageBuffer, mimeType)
-    } catch {}
-
-    if (!aiResult) {
-      return response.ok({ data: [] })
-    }
-
-    if (aiResult.crop === 'INVALID') {
-      return response.badRequest({ error: aiResult.instructions })
-    }
-
-    await user.load('farmerProfile')
-
-    const isHealthy = aiResult.disease === 'HEALTHY'
-
-    await CropScan.create({
-      farmer_profile_id: user.farmerProfile!.id,
-      crop: aiResult.crop,
-      disease: isHealthy ? null : aiResult.disease,
-      instructions: aiResult.instructions,
-      search_term: isHealthy ? null : aiResult.search_term,
-      active_ingredient: isHealthy ? null : aiResult.active_ingredient,
-      category: isHealthy ? null : (aiResult.category as ProductCategory),
-    })
-
-    if (aiResult.disease === 'HEALTHY') {
-      return response.ok({
-        data: {
-          diagnosis: {
-            instructions: aiResult.instructions,
-            crop: aiResult.crop,
-          },
-        },
-      })
-    }
-
-    const result = await getCropTreatmentResults(aiResult)
-
-    return response.ok({
-      data: {
-        diagnosis: {
-          crop: aiResult.crop,
-          disease: aiResult.disease,
-          instructions: aiResult.instructions,
-        },
-        treatments:
-          result?.rows && result.rows.length
-            ? result.rows.map((row) => ({
-                ...row,
-                links: {
-                  create_order: {
-                    method: 'POST',
-                    href: router.makeUrl('api.v1.products.orders.store', [row.id]),
-                  },
-                },
-              }))
-            : [],
+        /**
+         * @todo: Decide whether to return the crop scan store endpoint in links.
+         */
       },
     })
   }
