+++ start/routes/api/v1/farmer_profile_routes.ts
@@ -10,15 +10,6 @@ router
       .apiOnly()
       .only(['store', 'show'])
       .middleware('show', [middleware.auth(), middleware.role({ role: UserRolesEnum.Farmer })])
-
-    // Route for farmer to scan a crop and get treatment results.
-    router
-      .post('farmer_profiles/:farmer_profile_id/diagnose', [
-        () => import('#controllers/farmer_profiles_controller'),
-        'diagnose',
-      ])
-      .as('farmer_profiles.diagnose')
-      .use([middleware.auth(), middleware.role({ role: UserRolesEnum.Farmer })])
   })
   .prefix('api/v1')
   .as('api.v1')
